+++ app/helpers/claims_helper.rb
@@ -46,12 +46,13 @@ module ClaimsHelper
   end
 
   def payment_answers(claim)
+    change_slug = claim.building_society? ? "building-society-account" : "personal-bank-account"
     [].tap do |a|
       a << [translate("questions.bank_or_building_society"), claim.bank_or_building_society.to_s.humanize, "bank-or-building-society"]
-      a << ["Name on bank account", claim.banking_name, "bank-details"]
-      a << ["Bank sort code", claim.bank_sort_code, "bank-details"]
-      a << ["Bank account number", claim.bank_account_number, "bank-details"]
-      a << ["Building society roll number", claim.building_society_roll_number, "bank-details"] if claim.building_society_roll_number.present?
+      a << ["Name on bank account", claim.banking_name, change_slug]
+      a << ["Bank sort code", claim.bank_sort_code, change_slug]
+      a << ["Bank account number", claim.bank_account_number, change_slug]
+      a << ["Building society roll number", claim.building_society_roll_number, change_slug] if claim.building_society_roll_number.present?
     end
   end