+++ spec/models/student_loans/slug_sequence_spec.rb
@@ -50,5 +50,21 @@ RSpec.describe StudentLoans::SlugSequence do
         expect(slug_sequence.slugs).to include("student-loan-start-date")
       end
     end
+
+    context "when claim payment details are 'personal bank account'" do
+      it "excludes the 'building-society-account' slug" do
+        claim.bank_or_building_society = :personal_bank_account
+
+        expect(slug_sequence.slugs).not_to include("building-society-account")
+      end
+    end
+
+    context "when claim payment details are 'building society'" do
+      it "excludes the 'personal-bank-account' slug" do
+        claim.bank_or_building_society = :building_society
+
+        expect(slug_sequence.slugs).not_to include("personal-bank-account")
+      end
+    end
   end
 end